Clarifying PR requirements
Digital Ocean has cracked down on low quality PRs in recent years. On the PR dashboard, there's a warning against ineligible repos.
> **Ineligible Repository**
Your PR was submitted to a repository that is not participating in Hacktoberfest. Maintainers of the repository can add the "hacktoberfest" topic to their repository if they wish to participate.

I've clarified the page to add a link to what is considered an elegible repository.
The original `Requirements: 4 pull requests in **any** repository` is disingenuous